requirements.txt
# Core dependencies
python-telegram-bot>=21.0.0
twscrape>=0.17.0
redis>=5.0.1
asyncpg>=0.29.0
asyncio-throttle>=1.0.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.2
python-dotenv>=1.0.0

# Web interface
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
aiohttp>=3.9.1
httpx>=0.25.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2